chore: refine Modal
